style(ruff): RUF 规则修复 + 配置完善（45 处）
ruff --fix --unsafe-fixes + 手动修复：
- RUF059 unused-unpacked-variable (35)：未使用解包变量改 _
- RUF013 implicit-optional (4)：x: str = None → x: str | None = None
- RUF005 collection-literal-concatenation (4)：list + [item] → [*list, item]
- RUF015 unnecessary-iterable-allocation-for-first-element (3)：list[0] → next(...)
- RUF022 unsorted-dunder-all (14→0)：项目惯例按子模块分组，全局 ignore
- RUF012 mutable-class-default (16→0)：类常量 UPPER_SNAKE_CASE 命名约定，全局 ignore
- RUF034 useless-if-else (1)：invoice.py 两分支同值，简化
- RUF046 unnecessary-cast-to-int (1)：int(round(x)) → round(x)
- RUF100 unused-noqa：清理无效 noqa

恢复 RUF100 误删的有效 noqa：
- multimodal/* (12)：per-file-ignores F401（可选依赖探测 import 副作用）
- mcp_server/server.py：import deadman 副作用 import
- ragas_evaluator.py：ragas.evaluate 探测 import
- durable_execution.py：嵌套 class 自引用 F821
- test_memorial_writer.py：WebServer 用 TYPE_CHECKING import
- test_p6_tool_auto_schema.py：恢复 Optional[str] 测试兼容用例 UP045 noqa

验证：compileall 通过，pytest 2617 passed。

+++ .traecli/src/deadman/plan_score/scorer.py
@@ -146,8 +146,8 @@ class PlanScorer:
         total = 0.0
         for cat, weight in WEIGHTS.items():
             total += score_map.get(cat, 0) * weight
-        # 四舍五入 + clamp
-        result = int(round(total))
+        # 四舍五入 + clamp（round 返回 int，无需 int() 包装）
+        result = round(total)
         if result < 0:
             return 0
         if result > 100:
